user/requirements.txt
#requirements.txt
boto3
langchain
streamlit
faiss-cpu
langchain_community
python-dotenv